Twilio

Twilio Monad

type Twilio = TwilioT IO

This monad allows you to make authenticated REST API requests to Twilio using your AccountSID and AuthToken.

runTwilio :: Credentials -> Twilio a -> IO a

Run zero or more REST API requests to Twilio.

runTwilio'

Arguments

:: IO String
Account SID

-> IO String
Authentication Token

-> Twilio a

-> IO a

Parse an account SID and authentication token before running zero or more REST API requests to Twilio.

For example, you can fetch the Calls resource in the IO monad as follows:

module Main where

import Control.Monad.IO.Class (liftIO)
import System.Environment (getEnv)
import Twilio.Calls as Calls
import Twilio.Types

-- | Print calls.
main :: IO ()
main = runTwilio' (getEnv "ACCOUNT_SID")
                  (getEnv "AUTH_TOKEN")
     $ Calls.get >>= liftIO . print

Credentials

type Credentials = (AccountSID, AuthToken)

Your AccountSID and AuthToken are used to make authenticated REST API requests to Twilio.

parseCredentials

Arguments

:: String
Account SID

-> String
Authentication Token

-> Maybe Credentials

Parse an account SID and authentication token to Credentials.

Account SID

data AccountSID

getSID :: SID s => s -> String

parseSID :: SID s => String -> Maybe s

Authentication Token

data AuthToken

Your authentication token is used to make authenticated REST API requests to your Twilio account.

getAuthToken :: AuthToken -> String

Get the String representation of an AuthToken.

parseAuthToken :: String -> Maybe AuthToken

Parse a String to an AuthToken.