Twilio.Call.Feedback

Description

Resource

data Feedback

Feedback is a subresource of a Call instance resource. It represents a call quality feedback entry for a given phone call.

Constructors

Feedback

Fields

sid :: !CallSID

accountSID :: !AccountSID

qualityScore :: !Quality

issues :: ![Issue]

dateCreated :: !UTCTime

dateUpdated :: !UTCTime

get :: MonadThrow m => CallSID -> TwilioT m Feedback

Get a Call's Feedback by CallSID.

Types

data Quality

An integer 1 to 5 quality score where 1 represents very poor call quality and 5 represents a perfect call.

Constructors

Q1
Very poor call quality

Q2

Q3

Q4

Q5
A perfect call

data Issue

A list of issues experienced during the call.

Constructors

ImperfectAudio
Imperfect audio quality: Choppy, echoed, or garbled audio during conversation.

DroppedCall
Dropped call: call initially connected but was dropped.

IncorrectCallerId
Incorrect caller ID: Call connected but caller ID displayed 'Unknown' or an incorrect number.

PostDialDelay
Post dial delay: Call connected but there was a long delay between dialing the phone number and the start of ringing.

DigitsNotCaptured
DTMF tones not captured: Failed to capture digit input on phone menus.

UnsolicitedCall
Unsolicited call: Received telemarketer, wrong number, automated or other type of unsolicited call.

AudioLatency
Audio latency: Call participants can hear each other but with significant audio delay.

OneWayAudio
One way audio: Only one party could hear the audio during the conversation.