Twilio.Calls.FeedbackSummary

Resource

data FeedbackSummary

Constructors

FeedbackSummary

Fields

sid :: FeedbackSummarySID

startDate :: !Text

endDate :: !Text

accountSID :: !AccountSID

includeSubaccounts :: !Bool

status :: !FeedbackSummaryStatus

callCount :: !Int

callFeedbackCount :: !Int

qualityScoreAverage :: !Float

qualityScoreMedian :: !Int

qualityScoreStandardDeviation :: !Int

issues :: ![FeedbackSummaryIssue]

dateCreated :: !UTCTime

dateUpdated :: !UTCTime

data FeedbackSummaryIssue

Constructors

FeedbackSummaryIssue

Fields

description :: !Issue

count :: !Int

percentageOfTotalCalls :: !Float

Types

data FeedbackSummarySID

data Issue

A list of issues experienced during the call.

Constructors

ImperfectAudio
Imperfect audio quality: Choppy, echoed, or garbled audio during conversation.

DroppedCall
Dropped call: call initially connected but was dropped.

IncorrectCallerId
Incorrect caller ID: Call connected but caller ID displayed 'Unknown' or an incorrect number.

PostDialDelay
Post dial delay: Call connected but there was a long delay between dialing the phone number and the start of ringing.

DigitsNotCaptured
DTMF tones not captured: Failed to capture digit input on phone menus.

UnsolicitedCall
Unsolicited call: Received telemarketer, wrong number, automated or other type of unsolicited call.

AudioLatency
Audio latency: Call participants can hear each other but with significant audio delay.

OneWayAudio
One way audio: Only one party could hear the audio during the conversation.