Twilio.Internal.Parser

Documentation

(<&>) :: Functor f => f a -> (a -> b) -> f b

maybeReturn' :: Maybe (Maybe a) -> Parser (Maybe a)

maybeReturn'' :: Maybe (Maybe (Maybe a)) -> Parser (Maybe a)

filterEmpty :: Text -> Maybe Text

parseDate :: (Monad m, MonadPlus m) => Text -> m UTCTime

parseDateTime :: (Monad m, MonadPlus m) => Text -> m UTCTime

maybeReturn :: (Monad m, MonadPlus m) => Maybe a -> m a

newtype NonEmptyText

Constructors

NonEmptyText

Fields

getNonEmptyText :: Maybe Text

parseMaybeDateTime :: Maybe Text -> Parser (Maybe UTCTime)

Note that the parser only returns Nothing if the input is Nothing. If the input is an incorrectly formatted Text the parse will fail.

valueToText :: Value -> Maybe Text