Twilio.Types

Twilio Monad

type Twilio = TwilioT IO

This monad allows you to make authenticated REST API requests to Twilio using your AccountSID and AuthToken.

runTwilio :: Credentials -> Twilio a -> IO a

Run zero or more REST API requests to Twilio.

runTwilio'

Arguments

:: IO String
Account SID

-> IO String
Authentication Token

-> Twilio a

-> IO a

Parse an account SID and authentication token before running zero or more REST API requests to Twilio.

For example, you can fetch the Calls resource in the IO monad as follows:

module Main where

import Control.Monad.IO.Class (liftIO)
import System.Environment (getEnv)
import Twilio.Calls as Calls
import Twilio.Types

-- | Print calls.
main :: IO ()
main = runTwilio' (getEnv "ACCOUNT_SID")
                  (getEnv "AUTH_TOKEN")
     $ Calls.get >>= liftIO . print

Twilio Monad Transformer

data TwilioT m a

This monad transformer allows you to make authenticated REST API requests to Twilio using your AccountSID and AuthToken.

runTwilioT :: (MonadThrow m, MonadIO m) => Credentials -> TwilioT m a -> m a

Run zero or more REST API requests to Twilio, unwrapping the inner monad m.

runTwilioT'

Arguments

:: (MonadThrow m, MonadIO m)

=> m String
Account SID

-> m String
Authentication Token

-> TwilioT m a

-> m a

Parse an account SID and authentication token before running zero or more REST API requests to Twilio, unwrapping the inner monad m.

Making Requests

forAccount

Arguments

:: (MonadThrow m, MonadIO m)

=> AccountSID
Sub-Account SID

-> TwilioT m a
Zero or more Twilio REST API requests

-> TwilioT m a

Run zero or more Twilio REST API requests for an account that is a sub-account of your own.

For example, you can fetch the Calls resource for each of your account's sub-accounts as follows:

module Main where

import Control.Monad (forM_)
import Control.Monad.IO.Class (liftIO)
import System.Environment (getEnv)
import Twilio.Accounts as Accounts
import Twilio.Calls as Calls
import Twilio.Types

-- | Print calls for each sub-account.
main :: IO ()
main = runTwilio' (getEnv "ACCOUNT_SID")
                  (getEnv "AUTH_TOKEN") $ do
  subAccounts <- fmap getList Accounts.get
  forM_ subAccounts $ \Account {Accounts.sid = subAccountSID} ->
    forAccount subAccountSID Calls.get >>= liftIO . print

request :: (MonadIO m, FromJSON a) => String -> (Request -> Request) -> TwilioT m a

Given a relative resource URL, make an authenticated request to Twilio. You should never need to use this function directly.

requestForAccount :: (MonadIO m, FromJSON a) => String -> TwilioT m a

Given a relative resource URL, make an authenticated GET request to Twilio for the specified account. You should never need to use this function directly.

postForAccount :: (MonadIO m, FromJSON a) => String -> [(ByteString, ByteString)] -> TwilioT m a

Given a relative resource URL, make an authenticated POST request to Twilio for the specified account. You should never need to use this function directly.

Twilio Exceptions

data TwilioException

The set of Exceptions that may be thrown when attempting to make requests against Twilio's REST API.

Constructors

InvalidSID !String

InvalidAuthToken !String

InvalidCredentials

Credentials

type Credentials = (AccountSID, AuthToken)

Your AccountSID and AuthToken are used to make authenticated REST API requests to Twilio.

parseCredentials

Arguments

:: String
Account SID

-> String
Authentication Token

-> Maybe Credentials

Parse an account SID and authentication token to Credentials.

Authentication Token

data AuthToken

Your authentication token is used to make authenticated REST API requests to your Twilio account.

getAuthToken :: AuthToken -> String

Get the String representation of an AuthToken.

parseAuthToken :: String -> Maybe AuthToken

Parse a String to an AuthToken.

System Identifiers (SIDs)

class SID s where

Methods

getSID :: s -> String

parseSID :: String -> Maybe s

Resources

data AccountSID

data AddressSID

data ApplicationSID

data CallSID

data ConnectAppSID

data MessageSID

data PhoneNumberSID

data RecordingSID

data TranscriptionSID

data UsageTriggerSID

List Resources

class FromJSON b => List a b | a -> b where

Methods

getListWrapper :: Wrapper (Maybe PagingInformation -> [b] -> a)

Get the wrap-ed constructor of the List.

getList :: a -> [b]

The items in the List.

getPlural :: Const String (a, b)

The plural name for the items in the List.

parseJSONToList :: Value -> Parser a

Parse a JSON Value to an instance of the List.

data PagingInformation

Constructors

PagingInformation

Fields

pageNumber :: !Integer

The current page number. Zero-indexed, so the first page is 0.

numberOfPages :: !Integer

The total number of pages.

pageSize :: !Integer

How many items are in each page.

total :: !Integer

The total number of items in the list.

start :: !Integer

The position in the overall list of the first item in this page.

end :: !Integer

The position in the overall list of the last item in this page.

pageURI :: !(Maybe URI)

The URI of the current page.

firstPageURI :: !(Maybe URI)

The URI for the first page of this list.

nextPageURI :: !(Maybe URI)

The URI for the next page of this list.

previousPageURI :: !(Maybe URI)

The URI for the previous page of this list.

lastPageURI :: !(Maybe URI)

The URI for the last page of this list.

data AnsweredBy

Constructors

Human

Machine

data APIVersion

Constructors

API_2010_04_01

API_2008_08_01

data CallDirection

Constructors

Inbound

OutboundAPI

OutboundDial

data CallStatus

Constructors

Queued

Ringing

InProgress

Canceled

Completed

Failed

Busy

NoAnswer

data PriceUnit

Constructors

USD

EUR

JPY

OtherPriceUnit !String

data Wrapper a

wrap :: a -> Wrapper a

wraps a value so as not to break encapsulation.

(<&>) :: Functor f => f a -> (a -> b) -> f b

filterEmpty :: String -> Maybe String

parseDate :: (Monad m, MonadPlus m) => String -> m UTCTime

parseDateTime :: (Monad m, MonadPlus m) => String -> m UTCTime

parseMaybeDateTime :: Maybe String -> Parser (Maybe UTCTime)

Note that the parser only returns Nothing if the input is Nothing. If the input is an incorrectly formatted String the parse will fail.

safeRead :: (Monad m, MonadPlus m, Read a) => String -> m a

maybeReturn :: (Monad m, MonadPlus m) => Maybe a -> m a

maybeReturn' :: Maybe (Maybe a) -> Parser (Maybe a)

maybeReturn'' :: Maybe (Maybe (Maybe a)) -> Parser (Maybe a)

Misc

data ISOCountryCode

Country codes in ISO 3166-1 alpha-2 format supported by Twilio.

Constructors

AU
Australia

AT
Austria

BH
Bahrain

BE
Belgium

BR
Brazil

BG
Bulgaria

CA
Canada

CY
Cyprus

CZ
Czech Republic

DK
Denmark

DO
Dominican Republic

SV
El Salvador

EE
Estonia

FI
Finland

FR
France

GR
Greece

HK
Hong Kong

IE
Ireland

IL
Israel

IT
Italy

JP
Japan

LV
Latvia

LT
Lithuania

LU
Luxembourg

MT
Malta

MX
Mexico

NL
The Netherlands

NO
Norway

NZ
New Zealand

PE
Peru

PL
Poland

PT
Portugal

PR
Puerto Rico

RO
Romania

SK
Slovakia

ZA
South Africa

ES
Spain

SE
Sweden

CH
Switzerland

GB
United Kingdom

US
United States

data AddressRequirement

Constructors

None

Any

Local

Foreign

type Capabilities = Set Capability

data Capability

Constructors

Voice

SMS

MMS

parseCapabilitiesFromJSON :: Value -> Capabilities

data NonEmptyString