Twilio.Types.Alpha

Description

This module defines Alpha, a data type for the characters A through Z, along with a singleton type.

Documentation

data Alpha

Characters A through Z

Constructors

A

B

C

D

E

F

G

H

I

J

K

L

M

N

O

P

Q

R

S

T

U

V

W

X

Y

Z

alphaToChar :: Alpha -> Char

salphaToChar :: IsAlpha a => SAlpha a -> Char

data SAlpha a where

Constructors

SA :: SAlpha A

SB :: SAlpha B

SC :: SAlpha C

SD :: SAlpha D

SE :: SAlpha E

SF :: SAlpha F

SG :: SAlpha G

SH :: SAlpha H

SI :: SAlpha I

SJ :: SAlpha J

SK :: SAlpha K

SL :: SAlpha L

SM :: SAlpha M

SN :: SAlpha N

SO :: SAlpha O

SP :: SAlpha P

SQ :: SAlpha Q

SR :: SAlpha R

SS :: SAlpha S

ST :: SAlpha T

SU :: SAlpha U

SV :: SAlpha V

SW :: SAlpha W

SX :: SAlpha X

SY :: SAlpha Y

SZ :: SAlpha Z

class IsAlpha a where

Methods

promote :: SAlpha a

demote :: SAlpha a -> Alpha