Twilio.Types.List

Documentation

class FromJSON b => List a b | a -> b where

Minimal complete definition

getListWrapper, getList, getPlural

Methods

getListWrapper :: Wrapper (Maybe PagingInformation -> [b] -> a)

Get the wrap-ed constructor of the List.

getList :: a -> [b]

The items in the List.

getPlural :: Const Text (a, b)

The plural name for the items in the List.

parseJSONToList :: Value -> Parser a

Parse a JSON Value to an instance of the List.

data PagingInformation

Constructors

PagingInformation

Fields

pageNumber :: !Integer

The current page number. Zero-indexed, so the first page is 0.

numberOfPages :: !(Maybe Integer)

The total number of pages.

pageSize :: !Integer

How many items are in each page.

total :: !(Maybe Integer)

The total number of items in the list.

start :: !Integer

The position in the overall list of the first item in this page.

end :: !Integer

The position in the overall list of the last item in this page.

pageURI :: !(Maybe URI)

The URI of the current page.

firstPageURI :: !(Maybe URI)

The URI for the first page of this list.

nextPageURI :: !(Maybe URI)

The URI for the next page of this list.

previousPageURI :: !(Maybe URI)

The URI for the previous page of this list.

lastPageURI :: !(Maybe URI)

The URI for the last page of this list.

data Wrapper a

wrap :: a -> Wrapper a

wraps a value so as not to break encapsulation.