Control.Monad.Twilio

The Twilio monad

type Twilio = TwilioT IO

This monad allows you to make authenticated REST API requests to Twilio using your AccountSID and AuthToken.

runTwilio :: Credentials -> Twilio a -> IO a

Run zero or more REST API requests to Twilio.

runTwilio'

Arguments

:: IO String
Account SID

-> IO String
Authentication Token

-> Twilio a

-> IO a

Parse an AccountSID and AuthToken before running zero or more REST API requests to Twilio.

For example, you can fetch the Calls resource in the IO monad as follows:

module Main where

import Control.Monad.IO.Class (liftIO)
import System.Environment (getEnv)
import Twilio.Calls as Calls
import Twilio.Types

-- | Print calls.
main :: IO ()
main = runTwilio' (getEnv "ACCOUNT_SID")
                  (getEnv "AUTH_TOKEN")
     $ Calls.get >>= liftIO . print

The Twilio monad transformer

newtype TwilioT m a

This monad transformer allows you to make authenticated REST API requests to Twilio using your AccountSID and AuthToken.

Constructors

TwilioT (Monad m => (Credentials, AccountSID) -> RequestT m a)

runTwilioT :: MonadIO m => Credentials -> TwilioT m a -> m a

Run zero or more REST API requests to Twilio, unwrapping the inner monad m.

runTwilioT'

Arguments

:: (Functor m, MonadThrow m, MonadIO m)

=> m String
Account SID

-> m String
Authentication Token

-> TwilioT m a

-> m a

Parse an AccountSID and AuthToken before running zero or more REST API requests to Twilio, unwrapping the inner monad m.

Types

type Credentials = (AccountSID, AuthToken)

Your AccountSID and AuthToken are used to make authenticated REST API requests to Twilio.

data TwilioException

The set of Exceptions that may be thrown when attempting to make requests against Twilio's REST API.

Constructors

InvalidSID !Text

InvalidAuthToken !Text

InvalidCredentials

UnexpectedResponse !(Response ByteString)