Twilio.Internal.Request

Documentation

newtype RequestF a

RequestF represents an HTTP request and stores a continuaton for the eventual Response to the request.

Constructors

RequestF (Request, Response ByteString -> a)

newtype RequestT m a

RequestT m a augments an existing monad m with the ability to perform HTTP requests for Resources.

Constructors

RequestT

Fields

runRequestT :: FreeT RequestF m a

class Monad m => MonadRequest m where

Methods

request :: (Response ByteString -> m a) -> Request -> m a

baseURL :: Text

A dummy interpreter

runRequest' :: (Monad m, MonadIO m) => (Text, Text) -> RequestT m a -> m a